Bug 1241454: Fix enum class storage of ExprType; r=bustage

+++ js/src/asmjs/WasmTypes.h
@@ -145,7 +145,7 @@ class Val
 // void represented by the empty list). For now it's easier to have a flat enum
 // and be explicit about conversions to/from value types.
 
-enum class ExprType : uint8_t
+enum class ExprType : uint16_t
 {
     I32 = uint8_t(ValType::I32),
     I64 = uint8_t(ValType::I64),
